Move the stack-trace palette to where stack traces are coloured
`digression.core` depended on `iridescence` for two definitions in one
file: the eleven-colour `Palette` trait and its default instance, about
thirty lines describing how a rendered stack trace should look. Nothing
else in the component names a colour. Both move to `digression.ansi`,
which is where a stack trace is actually rendered and which already
depends on escapade and escritoire for that purpose.

The trait is renamed `StackTracePalette` because it cannot keep its old
name: as a member of `object StackTrace` it was `StackTrace.Palette`,
and a second `object StackTrace` in package `digression` is illegal,
while a bare `Palette` would collide with `iridescence.Palette`, which
the file imports.

The instance sits in `StackTracePalette`'s companion rather than at the
package level. That is not cosmetic: it used to resolve through
`StackTrace`'s companion, and a package-level given would have needed an
explicit import at every use site, since Scala 3's `*` does not import
givens. In the companion it stays in implicit scope and probably's two
`summon` sites keep working unchanged.

Cutting the edge exposed six more conduit users, each now declaring what
it uses: aviation, gesticulate, guillotine, urticose.url and
xylophone.core reached `contextual` through digression, and yossarian
reached `Chroma` — which is anticipation's opaque type, not one of
iridescence's — through the same path.

digression.core's closure falls from 26 modules to 19 and the critical
path from 26 to 25.
