+++ lib/digression/src/ansi/soundness_digression_ansi.scala
@@ -32,4 +32,5 @@
                                                                                                   */
 package soundness
 
-export digression.{exceptionTeletype, frameTeletype, methodTeletype, stackTraceTeletype}
+export digression.{exceptionTeletype, frameTeletype, methodTeletype, StackTracePalette,
+    stackTraceTeletype}
